Check out the output of the step line regex match
It was possible that a line which matched the initial regex did not
provide any output on the parsing regex. Without knowing what the
offending string was, I've added some defensive switches into the
parsing routine, so output might be a little different, but it shouldn't
throw and stop the build.

Change-type: patch
Signed-off-by: Cameron Diver <cameron@resin.io>